fixing checks warnings and errors

+++ blosc/context.h
@@ -41,7 +41,7 @@ struct blosc2_context_s {
   /* Flags for header */
   uint8_t blosc2_flags;
   /* Flags specific for blosc2 */
-  size_t sourcesize;
+  int32_t sourcesize;
   /* Number of bytes in source buffer */
   int32_t header_overhead;
   /* The number of bytes in chunk header */
@@ -53,9 +53,9 @@ struct blosc2_context_s {
   /* Length of the block in bytes */
   int32_t output_bytes;
   /* Counter for the number of input bytes */
-  size_t srcsize;
+  int32_t srcsize;
   /* Counter for the number of output bytes */
-  size_t destsize;
+  int32_t destsize;
   /* Maximum size for destination buffer */
   int32_t typesize;
   /* Type size */
